+++ webapp/src/pages/Dashboard.tsx
@@ -3,6 +3,7 @@ import noData from "assets/void.svg";
 import PerturbationTestingPreview from "components/Analysis/PerturbationTestingPreview";
 import PreviewCard from "components/Analysis/PreviewCard";
 import WarningsPreview from "components/Analysis/WarningsPreview";
+import Description from "components/Description";
 import Telescope from "components/Icons/Telescope";
 import Loading from "components/Loading";
 import PerformanceAnalysis from "components/Metrics/PerformanceAnalysis";
@@ -12,6 +13,8 @@ import React from "react";
 import { Link, useParams } from "react-router-dom";
 import { getDatasetInfoEndpoint } from "services/api";
 import { isPipelineSelected } from "utils/helpers";
+import { behavioralTestingDescription } from "./PerturbationTestingSummary";
+import { postprocessingDescription } from "./Threshold";
 
 const DEFAULT_PREVIEW_CONTENT_HEIGHT = 502;
 
@@ -43,7 +46,13 @@ const Dashboard = () => {
         paddingX={4}
         paddingY={1}
       >
-        <Typography variant="h2">Dashboard</Typography>
+        <Box display="flex" flexDirection="column">
+          <Typography variant="h2">Dashboard</Typography>
+          <Description
+            text="Explore the analyses of your datasets and models."
+            link="/"
+          />
+        </Box>
         <Button
           color="secondary"
           variant="contained"
@@ -59,6 +68,12 @@ const Dashboard = () => {
         <PreviewCard
           title="Dataset Class Distribution Analysis"
           to={`/${jobId}/dataset_class_distribution_analysis${searchString}`}
+          description={
+            <Description
+              text="Compare the class distribution of your training and evaluation sets."
+              link="/dataset-warnings/"
+            />
+          }
         >
           <Box height={DEFAULT_PREVIEW_CONTENT_HEIGHT}>
             <WarningsPreview jobId={jobId} />
@@ -66,7 +81,15 @@ const Dashboard = () => {
         </PreviewCard>
       )}
       {isPipelineSelected(pipeline) && (
-        <PreviewCard title="Performance Analysis">
+        <PreviewCard
+          title="Performance Analysis"
+          description={
+            <Description
+              text="Assess model performance through prediction metrics."
+              link="/#performance-analysis"
+            />
+          }
+        >
           <PerformanceAnalysis jobId={jobId} pipeline={pipeline} />
         </PreviewCard>
       )}
@@ -74,6 +97,7 @@ const Dashboard = () => {
         <PreviewCard
           title="Behavioral Testing"
           to={`/${jobId}/behavioral_testing_summary${searchString}`}
+          description={behavioralTestingDescription}
         >
           <Box height={DEFAULT_PREVIEW_CONTENT_HEIGHT}>
             <PerturbationTestingPreview
@@ -89,6 +113,7 @@ const Dashboard = () => {
           <PreviewCard
             title="Post-processing Analysis"
             to={`/${jobId}/thresholds${searchString}`}
+            description={postprocessingDescription}
           >
             <Box height={DEFAULT_PREVIEW_CONTENT_HEIGHT}>
               <ThresholdPlot jobId={jobId} pipeline={pipeline} />