+++ webapp/src/pages/PerturbationTestingSummary.tsx
@@ -1,4 +1,5 @@
 import { Box, Typography } from "@mui/material";
+import Description from "components/Description";
 import PerturbationTestingSummaryTable from "components/PerturbationTestingSummary/PerturbationTestingSummaryTable";
 import useQueryState from "hooks/useQueryState";
 import React from "react";
@@ -6,6 +7,13 @@ import { useParams } from "react-router-dom";
 import { getDatasetInfoEndpoint } from "services/api";
 import { PIPELINE_REQUIRED_TIP } from "utils/const";
 import { isPipelineSelected } from "utils/helpers";
+
+export const behavioralTestingDescription = (
+  <Description
+    text="Assess if your model is robust to small modifications."
+    link="/behavioral-testing-summary/"
+  />
+);
 
 const PerturbationTestingSummary = () => {
   const { jobId } = useParams<{ jobId: string }>();
@@ -16,9 +24,8 @@ const PerturbationTestingSummary = () => {
 
   return (
     <Box display="flex" flexDirection="column" height="100%">
-      <Typography variant="h4" paddingBottom={2}>
-        Behavioral Testing Summary
-      </Typography>
+      <Typography variant="h4">Behavioral Testing Summary</Typography>
+      {behavioralTestingDescription}
       {isPipelineSelected(pipeline) ? (
         <PerturbationTestingSummaryTable
           jobId={jobId}