+++ docs/docs/getting-started/changelog.md
@@ -1,6 +1,26 @@
 # Releases
 
-## [2.1.1] - 2021-06-06
+## [2.2.0] - 2022-07-08
+
+### Added
+
+- Contextual information has been added to most functionalities, linking directly to the relevant documentation.
+- The F1 metric was added as a default metric.
+
+### Changed
+
+- Smart tags are now grouped in families, so it is easier to digest them. This also impacts how filters are displayed.
+- The confusion matrix can be shown with raw values (not just normalized).
+- We now show the top 20, instead of 10, most similar examples in the utterances details.
+- Renamed our docker images to use `servicenowdocker` registry on Docker Hub.
+
+##### Performance Overview Table
+- Columns can be temporarily hidden.
+- Use outcomes' icons as headers instead of the outcomes' full names.
+- Added visual bars, so it is easier to analyze the model's performance.
+
+
+## [2.1.1] - 2022-06-06
 
 ### Changed
 
@@ -19,7 +39,7 @@
 - Fixed unexpected borders in Performance Analysis table.
 - Fixed proposed actions which couldn't be applied on an utterance in the latest release.
 
-## [2.1.0] - 2021-05-27
+## [2.1.0] - 2022-05-27
 
 - Ability to get predictions without postprocessing in the exploration space. See section "Excluding
   Post-Processing" [here](../user-guide/exploration-space/index.md).
@@ -28,6 +48,6 @@
 - Links on top words to filter utterances that contain it. See the section "Word
   Clouds" [here](../user-guide/exploration-space/performance-overview.md).
 
-## [2.0.0] - 2021-04-12
+## [2.0.0] - 2022-04-12
 
 First public release.